requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
